Centralize collection of crate metadata
Previously, crate metadata — specifically, name, version, and homepage
URL — was retrieved (with the `env!` macro) wherever it was about to
be used — specifically, when reporting this metadata to an IRC user on
request, when constructing a default realname, and when constructing a
default quit message).

Furthermore, this information was gathered inconsistently;
specifically, when reporting it to an IRC user on request, the version
information reported by `git describe` would be used if available, but
otherwise only the version information reported by Cargo would be
used.

This change centralizes the collection of such crate metadata and
adds methods to the `State` type for retrieving such metadata, with
the hope of improving consistency and DRY-ness.