Bump version to 0.2.0
Bump the crate version to 0.2.0 in preparation for a release to
Crates.io.

I don't think the code really deserves such a release at this point,
but I may not have time to work on it for a long while, and it's much
better than what I have on Crates.io now.