Upgrade to Rust 1.24.1
Bump the Rust toolchain version specified in `shell.nix` from 1.22.1
to 1.24.1 to escape a `jobserver` bug:
<https://github.com/rust-lang/rust/issues/46981>.

Version 1.24.1 comes with rustfmt, so this patch also drops the
dependency on the Nix package for rustfmt. The version of rustfmt
bundled with rustc 1.24.1 does not accept any of the `rustfmt.toml`
configuration file options I was using, so this patch deletes
`rustfmt.toml`.