Update to Rust 1.29.1 for security fix
This patch updates the version of Rust with which I build from 1.26.2
to 1.29.1, because the latter includes a fix for a security
vulnerability.

This patch also removes a linting attribute that disabled the
`unused_doc_comment` lint. I had been disabling this lint because
`error-chain`'s macro `error_chain!` was triggering it, which no
longer appears to be the case.