Combine generic module feature handler arguments
Combine the `&State` and `MsgMetadata` arguments passed to handlers
for both bot commands and triggers, as well as information on the
feature that the handler is to handle, into a single `HandlerContext`
structure. Giving handlers a structure that knows all this information
unlocks various API improvements I have among my agenda.