Replace read_aatxe_client with with_aatxe_client
The GitHub user @ctm pointed out in [GitHub issue #52] that the bot
module `quote` relied on the `pub(crate)` method
`State::read_aatxe_client`, which allows access to `IrcClient`s from
the `irc` crate. I would prefer that `quote` not rely on anything not
in the public API; however, I also would not want to fix the `irc`
crate as our IRC backend by exposing it in the public API. This patch
compromises between these motivations by exposing the functionality in
question behind a feature flag, `aatxe-irc`.

Also, the name `read_aatxe_client` seemed unfortunate, because it
implies that the function returns a lock read guard as do the other
`read_*` methods of `State`, when it does not. This patch renames the
function to `with_aatxe_client` and, additionally, documents it.

GitHub: resolve #52

[GitHub issue #52]: <https://github.com/8573/irc-bot.rs/issues/52>