Config: Correctly place channel quoting warning
In commit ee3d54202ebe8325b9327ff295cf84614920cddb, I added to the
configuration documentation a warning that one should quote channel
names lest their leading '#' characters cause them otherwise to be
interpreted as YAML comments.  Unfortunately, I incorrectly placed
this warning in the documentation for the `name` field for *servers*
rather than for channels.  This patch moves the warning to the
documentation of the `name` field for channels.

GitHub: Resolve #58